Renamed QueryTable.Name to QueryTable.TableName
As the DSL functionality encourages creating new classes deriving from
QueryTable and adding fields on the new classes that represents the
database fields of a specific table, a property named "Name" is not very
clever on the base class, as there is a high probability that there is a
database field called "Name". Therefore the property was renamed to
TableName, as there is a higher probability that there is not a field
named this in the table.